Add more challenges
Per issue #1

Added in more challenges for the bingo card since `BingoCard` will
not overflow with more items to use since it relies on `slice()`
to populate the table.

This makes it possible to add more items which will still be
shuffled, meaning new challenges can be added by modifying
`App.data.tasks`.